+++ cmd/bbp-scope/exclude.txt
